+++ src/util/frame/side_data.rs
@@ -219,7 +219,7 @@ impl<'a> SideData<'a> {
 
     #[inline]
     pub fn data(&self) -> &[u8] {
-        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size) }
+        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size as usize) }
     }
 
     #[inline]
